Reorganize filter components, replace Icon with MUI
