+++ app/assets/stylesheets/modules/mobile-nav.scss
@@ -102,7 +102,6 @@
 
 .filter-btn,
 .coins-btn {
-  width: 180px;
   text-transform: none;
 }
